adjusted tests for moved endpoint
