+++ Pinionszek_API.Tests/Tests/UnitTests/ApiServices/PaymentApiServiceTests.cs
@@ -235,5 +235,23 @@ namespace Pinionszek_API.Tests.Tests.UnitTests.ApiServices
             payment6?.SharedPayment.Should().NotBeNull();
             payment6?.SharedPayment?.IdSharedPayment.Should().BeGreaterThan(0);
         }
+
+        [Fact]
+        public async Task PaymentApiService_GetDefaultGeneralCategoriesAsync_ReturnsCategoriesOrNotfound()
+        {
+            //Arrange
+            var dbContext = new InMemContext().GetDatabaseContext();
+            var paymentApiService = new PaymentApiService(await dbContext);
+
+            //Act
+            var defaultCategories = await paymentApiService.GetDefaultGeneralCategoriesAsync();
+
+            //Assert
+            defaultCategories.Should().NotBeNullOrEmpty();
+            defaultCategories.Count().Should().Be(3);
+            defaultCategories
+                .Where(dc => string.IsNullOrEmpty(dc.Name))
+                .Should().BeNullOrEmpty();
+        }
     }
 }
